[libfko] changed 'state' context element to 'int' type to fix a 'extra high-order bits' bug found by Coverity

+++ lib/fko_context.h
@@ -94,7 +94,7 @@ struct fko_context {
     int             added_gpg_prefix;
 
     /* State info */
-    unsigned short  state;
+    unsigned int    state;
     unsigned char   initval;
 
 #if HAVE_LIBGPGME
